requirements.txt
alembic==1.7.4
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
Flask==1.1.4
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==0.15.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.5
markdown2==2.4.1
MarkupSafe==2.0.1
Pillow==8.4.0
psycopg2==2.9.1
requests==2.26.0
SQLAlchemy==1.4.26
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.3.3
zipp==3.6.0
